requirements.txt
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Cors
Python-Dotenv
PyMySQL